Move service to separate module
